+++ integration-test/list_test.go
@@ -29,7 +29,7 @@ func (s *RunVSuite) TestListSleep(c *check.C) {
 
 	go func() {
 		defer close(exitChan)
-		_, exitCode := s.runvCommand(c, "start", "--bundle", s.bundlePath, ctrName)
+		_, exitCode := s.runvCommand(c, "run", "--bundle", s.bundlePath, ctrName)
 		c.Assert(exitCode, checker.Equals, 0)
 	}()
 
@@ -56,7 +56,7 @@ func (s *RunVSuite) TestListSleepJson(c *check.C) {
 
 	go func() {
 		defer close(exitChan)
-		_, exitCode := s.runvCommand(c, "start", "--bundle", s.bundlePath, ctrName)
+		_, exitCode := s.runvCommand(c, "run", "--bundle", s.bundlePath, ctrName)
 		c.Assert(exitCode, checker.Equals, 0)
 	}()
